Question 3 Add a new funtion to this library's API.
The function should search a TLV encoded byte array for a specified
"tag", and return the byte offset of the tag with the array.

Add: tlv_search() is based on tlv_parse() but returns as soon as it
finds the specified tag, or TLV_ERR_ENOENT if the tag is not found.
TLVToken *t, was removed, as it was not specified to parse the TLV
objects.

[in/out] nTok    Input size of the array / Output number of tokens parsed
[input]  src     Source pointer to TLV data
[input]  srcLen  Length source data to parse
[input]  type    Type of tag to find

Returns the byte offset of the first tag with the specified type, or negative error

int tlv_search(int* nTok, const void* src, int srcLen, int type);